Export dev config du fichier boot
